CRUD update

+++ src/main/java/com/sprint/BookPartnerApplication/repository/JobsRepository.java
@@ -3,6 +3,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.sprint.BookPartnerApplication.entity.Jobs;
 @Repository
-public interface JobsRepository extends JpaRepository<Jobs, Long> {
+public interface JobsRepository extends JpaRepository<Jobs, Short> {
 
 }
